# openapi.yaml
---
openapi: 3.1.0
paths:
  /api/auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "400":
          description: Bad Request
      summary: Login
      tags:
      - Auth Resource
  /api/auth/me:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Current User
      tags:
      - Auth Resource
      security:
      - SecurityScheme: []
  /api/auth/register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "400":
          description: Bad Request
      summary: Register
      tags:
      - Auth Resource
  /api/auth/switch-org:
    post:
      parameters:
      - name: orgId
        in: query
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Switch Organization
      tags:
      - Auth Resource
      security:
      - SecurityScheme: []
  /api/blocks:
    get:
      parameters:
      - name: pageId
        in: query
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Blocks By Page
      tags:
      - Block Resource
      security:
      - SecurityScheme: []
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/JsonValue"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create Block
      tags:
      - Block Resource
      security:
      - SecurityScheme: []
  /api/blocks/batch:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/JsonValue"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Batch Save Blocks
      tags:
      - Block Resource
      security:
      - SecurityScheme: []
  /api/blocks/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/JsonValue"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update Block
      tags:
      - Block Resource
      security:
      - SecurityScheme: []
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete Block
      tags:
      - Block Resource
      security:
      - SecurityScheme: []
  /api/health:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Health
      tags:
      - Health Resource
  /api/org:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Current Org
      tags:
      - Org Resource
      security:
      - SecurityScheme: []
  /api/org/whoami:
    get:
      parameters:
      - name: Host
        in: header
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Whoami
      tags:
      - Org Echo Resource
  /api/pages:
    get:
      parameters:
      - name: workspaceId
        in: query
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List Pages
      tags:
      - Page Resource
      security:
      - SecurityScheme: []
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create Page
      tags:
      - Page Resource
      security:
      - SecurityScheme: []
  /api/pages/{id}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete Page
      tags:
      - Page Resource
      security:
      - SecurityScheme: []
  /api/whoami:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Who Am I
      tags:
      - Who Am I Resource
  /api/whoami/me:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Me
      tags:
      - Who Am I Resource
      security:
      - SecurityScheme: []
  /api/workspaces:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List Workspaces
      tags:
      - Workspace Resource
      security:
      - SecurityScheme: []
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create Workspace
      tags:
      - Workspace Resource
      security:
      - SecurityScheme: []
components:
  schemas:
    JsonValue:
      type: object
      properties:
        valueType:
          $ref: "#/components/schemas/ValueType"
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          pattern: \S
        password:
          type: string
          pattern: \S
    PageCreateRequest:
      type: object
      properties:
        workspaceId:
          type: integer
          format: int64
        title:
          type: string
        iconEmoji:
          type: string
        parentPageId:
          type: integer
          format: int64
    RegisterRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          pattern: \S
        password:
          type: string
          minLength: 8
          pattern: \S
        name:
          type: string
    ValueType:
      type: string
      enum:
      - ARRAY
      - OBJECT
      - STRING
      - NUMBER
      - "TRUE"
      - "FALSE"
      - "NULL"
    WorkspaceCreateRequest:
      type: object
      properties:
        name:
          type: string
        iconEmoji:
          type: string
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
      description: Authentication
info:
  title: vno-backend API
  version: 0.0.1